bot/requirements.txt
aiogram>=3.13.1
python-dotenv>=1.0.1
sqlalchemy>=2.0.35
asyncpg>=0.29.0
alembic>=1.13.0
loguru>=0.7.2
